Role is not correct
Below article is not correct because the Cognitive Services Contributor does not have the permission to manage Microsoft.Network 

------
https://docs.microsoft.com/en-us/azure/cognitive-services/cognitive-services-virtual-networks?tabs=portal#required-permissions
Required permissions
To apply a virtual network rule to a Cognitive Services resource, the user must have the appropriate permissions for the subnets being added. The required permission is the default Contributor role, or the Cognitive Services Contributor role. Required permissions can also be added to custom role definitions.
-------

https://docs.microsoft.com/en-us/azure/role-based-access-control/built-in-roles#cognitive-services-contributor